backend/requirements.txt
flask 
flask_socketio
GitPython 
unstructured

langchain
langchain-core
langchain-community
langchain_experimental

langchain-openai
openai  

langchain-anthropic

langchain-google-genai
google-generativeai

langchain-huggingface
transformers
huggingface-hub

python-dotenv

numpy
scikit-learn

youtube-transcript-api
faiss-cpu
tiktoken
